+++ examples/openapi/src/main.rs
@@ -137,7 +137,7 @@ mod tests {
             "openapi": "3.1.0",
             "info": {
                 "title": "Nidus Example API",
-                "version": "1.0.1"
+                "version": "1.0.2"
             },
             "paths": {
                 "/users": {